+++ authentication/aws/aws.go
@@ -1,4 +1,4 @@
-package aws_authentication
+package aws
 
 import (
 	"github.com/aws/aws-sdk-go/aws"
